+++ lib/datamappify/repository/query_method/method/reference_handler.rb
@@ -27,9 +27,13 @@ module Datamappify
           # @return [void]
           def perform_write
             referenced_entities.each do |entity|
-              entity.send("#{reference_key}=", @entity.id)
-              @repository.states.mark_as_dirty(entity)
-              @repository.save!(entity)
+              if entity.destroy?
+                @repository.destroy!(entity)
+              else
+                entity.send("#{reference_key}=", @entity.id)
+                @repository.states.mark_as_dirty(entity)
+                @repository.save!(entity)
+              end
             end
           end